--- actions/cherry-picker/functions/gh_auth_login.py
@@ -1,8 +1,0 @@
-# import subprocess
-
-# def gh_auth_login():
-
-#     print("Loggin!!!")
-    
-#     subprocess.run(['echo', 'config'])
-
